agent/pyproject.toml
[tool.poetry]
name = "mcp-agent"
version = "0.1.0"
description = "Starter"
authors = ["Markus Ecker <markus.ecker@gmail.com>"]
license = "MIT"
packages = [{include = "mcp_agent"}]

[tool.poetry.group.dev.dependencies]
autoflake = "^2.3.1"

[project]
name = "mcp-agent"
version = "0.1.0"
dependencies = [
    "langchain-openai>=0.2.1",
    "langchain-anthropic>=0.2.1",
    "langchain>=0.3.1",
    "openai>=1.51.0",
    "langchain-community>=0.3.1",
    "copilotkit==0.1.39",
    "uvicorn>=0.31.0",
    "python-dotenv>=1.0.1",
    "langchain-core>=0.3.25",
    "langgraph-cli[inmem]>=0.1.64",
    "langchain-mcp-adapters>=0.0.3",
    "fastmcp>=0.4.1",
    "langgraph (>=0.3.31,<0.4.0)",
    "langgraph-checkpoint (>=2.0.24,<3.0.0)",
    "sentry-sdk[fastapi] (>=2.26.1,<3.0.0)",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.0.1",
    "requests (>=2.32.3,<3.0.0)",
    "redis>=5.0.1",
    "openai-agents>=0.0.13"
]

[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[tool.poetry.dependencies]
python = ">=3.10,<3.14"
langchain-openai = "^0.2.1"
langchain-anthropic = "^0.2.1"
langchain = "^0.3.1"
openai = "^1.51.0"
langchain-community = "^0.3.1"
copilotkit = "0.1.39"
uvicorn = "^0.31.0"
python-dotenv = "^1.0.1"
langchain-core = "^0.3.25"
langgraph-cli = {extras = ["inmem"], version = "^0.1.64"}
langchain-mcp-adapters = "^0.0.3"
fastmcp = "^0.4.1"
langgraph = "^0.3.5"
litellm = "^1.43.12"
mem0ai = "^0.1.17"
pydantic = "^2.8.2"
logfire = "^0.19.0"
psycopg2-binary = "^2.9.9"
redis = "^5.0.1"
openai-agents = ">=0.0.13"

[tool.poetry.scripts]
demo = "mcp_agent.run_server:main"
custom-server = "mcp_agent.custom_server_entry:run_custom_server"

